--- public/starter-code/public/new.html
@@ -1,94 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Blog</title>
-    <link rel="stylesheet" href="vendor/styles/normalize.css">
-    <link rel="stylesheet" href="styles/icons.css">
-    <link rel="stylesheet" href="styles/base.css">
-    <link rel="stylesheet" href="styles/layout.css">
-    <link rel="stylesheet" href="styles/modules.css">
-    <link rel="stylesheet" href="vendor/styles/railscasts.css">
-    <link rel="icon" type="image/x-icon" href="https://avatars2.githubusercontent.com/u/18607604?v=3&s=200">
-
-    <script id="article-template" type="text/x-handlebars-template">
-      <article
-        class="clearfix"
-        data-author="{{authorSlug}}"
-        data-category="{{categorySlug}}">
-        <header>
-          <h1>{{title}}</h1>
-          <div class="byline">
-            By <address><a href="{{authorUrl}}">{{author}}</a></address>
-            {{publishStatus}}
-          </div>
-        </header>
-        <section class="article-body">{{{body}}}</section>
-        <a href="#" class="read-on">Read on &rarr;</a>
-      </article>
-    </script>
-    </head>
-  <body>
-    <header class="site-header clearfix">
-      <section class="heading-group-h1 pull-left">
-        <h1>
-          Kilovolt Blog!
-        </h1>
-        <h2>The shocking power of static pages</h2>
-      </section>
-      <nav class="main-nav">
-        <div class="icon-menu"></div>
-        <ul>
-          <li class="tab" data-content="write"><a href="#" class="icon-home"> Write</a></li>
-          <li class="tab" data-content="articles"><a href="#" class="icon-home"> Preview</a></li>
-        </ul>
-      </nav>
-    </header>
-    <main class="clearfix">
-      <section id="write" class="tab-content">
-        <h1>New Article</h1>
-        <form action="#" id="new-form">
-          <label>
-            <input type="text" id="article-title" placeholder="Article title" required>
-          </label>
-          <textarea id="article-body" rows="8" cols="40" required></textarea>
-          <label>
-            <input type="text" id="article-author" placeholder="Author Name" required>
-          </label>
-          <label>
-            <input type="text" id="article-author-url" placeholder="Author's URL" required>
-          </label>
-          <label>
-            <input type="text" id="article-category" placeholder="Category" required>
-          </label>
-          <label>
-            <input type="checkbox" id="article-published"> Published
-          </label>
-        </form>
-      </section>
-      <section id="articles" class="tab-content">
-
-      </section>
-      <section id="article-export">
-        <label id="export-field" for="article-json" class="pull-right">
-          <input type="text" readonly id="article-json">
-          When finalized, copy the above JSON, and paste into blogArticles.js
-        </label>
-      </section>
-    </main>
-    <footer>
-      &copy; 2017 Code Fellows | Static Blog | Happy Ipsum!
-    </footer>
-
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/4.0.4/handlebars.js"></script>
-    <script src="vendor/scripts/marked.js"></script>
-    <script src="vendor/scripts/highlight.pack.js"></script>
-    <script src="scripts/article.js"></script>
-    <script src="scripts/articleView.js"></script>
-    <script>
-      articleView.initNewArticlePage();
-    </script>
-  </body>
-</html>
