--- public/starter-code/public/styles/modules.css
@@ -1,83 +1,0 @@
-.site-header {
-  background: #FFEAAA;
-}
-
-.heading-group-h1 {
-  float: left;
-  padding: 0 0 .2em 1em;
-}
-
-.icon-menu {
-  font-size: 3em;
-  float: right;
-}
-
-.main-nav {
-  padding: 0.1em;
-}
-
-.main-nav ul {
-  display: none;
-  padding: 0;
-  list-style: none;
-  clear: both;
-}
-
-.main-nav:hover ul {
-  display: block;
-}
-
-.main-nav li {
-  margin-right: .4em;
-  font-size: 1.5em;
-  padding: .5em .2em;
-  border-bottom: 1px solid #ccc;
-}
-
-#filters {
-  padding: 0;
-  list-style: none;
-}
-
-.template {
-  display: none;
-}
-
-.byline {
-  color: #777;
-}
-
-.byline a {
-  font-weight: bold;
-  text-decoration: none;
-}
-
-#article-export label {
-  font-size: .8em;
-  font-style: italic;
-  color: #ccc;
-}
-
-@media only screen and (min-width: 640px) {
-  .icon-menu {
-    display: none;
-  }
-
-  .main-nav {
-    padding: .2em .2em .2em 2em;
-  }
-
-  .main-nav ul {
-    display: block;
-  }
-
-  .main-nav ul li {
-    display: inline;
-    border-bottom: none;
-  }
-
-  .main-nav ul li:hover a {
-    color: #29506D;
-  }
-
-}
